Refactor editors to share components
